requirements.txt
Flask>=2.2.0
Flask-Login>=0.6.2
Flask-SQLAlchemy>=3.0.2
stripe>=5.0.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.6
gunicorn>=20.1.0
Werkzeug~=3.1.3